requirements.txt
fastapi==0.110.0
httpx==0.27.0
uvicorn==0.30.1
PyYAML==6.0.2

